+++ src/game/main/HUD.java
@@ -4,7 +4,7 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int health = 100;
+    public static float health = 100;
 
     private short   redValue = 0,
             greenValue = 255,
@@ -32,7 +32,7 @@ public class HUD {
         g.setColor(Color.gray);
         g.fillRect(15,15, 200, 32);
         g.setColor(new Color(redValue, greenValue, blueValue));
-        g.fillRect(15,15, ( health * 2 ), 32);
+        g.fillRect(15,15, (int) (health * 2 ), 32);
         g.setColor(Color.white);
         g.drawRect(15,15,200,32);
